AdminPanel/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.0.1
billiard==4.1.0
binance-connector==3.3.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
cssbeautifier==1.14.9
daphne==4.0.0
Django==4.2.4
django-debug-toolbar==4.2.0
django-formtools==2.4.1
django-otp==1.2.2
django-phonenumber-field==6.4.0
django-rest-framework==0.1.0
django-rosetta==0.9.9
django-two-factor-auth==1.15.4
djangorestframework==3.14.0
djlint==1.32.1
EditorConfig==0.12.3
et-xmlfile==1.1.0
greenlet==2.0.2
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsbeautifier==1.14.9
json5==0.9.14
jwt==1.3.1
kombu==5.3.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
openpyxl==3.1.2
packaging==23.1
paramiko==3.3.1
pathspec==0.11.2
phonenumbers==8.13.19
polib==1.2.0
prompt-toolkit==3.0.39
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.8.8
requests==2.31.0
rest-framework-simplejwt==0.0.2
sendgrid==6.10.0
service-identity==23.1.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tomli==2.0.1
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.2
zope.interface==6.0
